# docs/OpenApi.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestión Ganadera - API Docs
  version: 0.0.1
  description: >-
    Documentación interactiva de la API REST para la gestión y trazabilidad
    ganadera de la empresa. Permite a los desarrolladores explorar y probar los
    endpoints de la API, facilitando la integración con sistemas externos y
    aplicaciones móviles.
servers:
  - url: 'http://localhost:8000/api'
...
security:
  - http: []
paths:
...
  /v1/animales:
    get:
      operationId: animales.index
      description: >-
        Retorna el listado paginado de animales registrados en el sistema con
        soporte de filtros y ordenamiento.
      summary: Listar animales
      tags:
        - Animal
      responses:
        '200':
          description: Paginated set of `AnimalResource`
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AnimalCollection'
                  message:
                    type: string
                    const: Listado de animales obtenido correctamente.
                required:
                  - data
                  - message
      security: []
    post:
      operationId: animales.store
      description: >-
        Valida los datos recibidos y registra un nuevo animal en el sistema.

        Valida la capacidad máxima del potrero asignado antes de completar el
        registro.
      summary: Registrar un nuevo animal
      tags:
        - Animal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreAnimalRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Animal creado correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/AnimalResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
  '/v1/animales/{id}':
    get:
      operationId: animales.show
      description: >-
        Retorna la información completa de un animal específico por su ID,

        incluyendo sus relaciones con raza, potrero, historial de pesajes y
        tratamientos.
      summary: Obtener detalle de un animal
      tags:
        - Animal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del animal.
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Animal obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/AnimalResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: animales.update
      description: >-
        Actualiza la información de un animal existente por su ID.

        Si se modifica la asignación de potrero, valida que el nuevo potrero
        tenga capacidad disponible.
      summary: Actualizar datos de un animal
      tags:
        - Animal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del animal a actualizar.
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnimalRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Animal actualizado correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/AnimalResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: animales.destroy
      description: >-
        Elimina el registro de un animal del sistema por su ID.

        Impide la eliminación si el animal posee registros de pesaje en su
        historial.
      summary: Eliminar un animal
      tags:
        - Animal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del animal a eliminar.
          schema:
            type:
              - integer
              - string
      responses:
        '204':
          description: No content
      security: []
  '/v1/potreros/{potrero}/animales':
    get:
      operationId: potreros.animales.index
      description: >-
        Retorna el listado paginado de animales asignados a un potrero
        específico.

        Valida previamente la existencia del potrero indicado en la ruta.
      summary: Listar animales por potrero
      tags:
        - Animal
      parameters:
        - name: potrero
          in: path
          required: true
          description: Identificador único del potrero.
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Listado de animales del potrero obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - type: object
                        additionalProperties:
                          type: string
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                      last_page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
                    required:
                      - current_page
                      - last_page
                      - per_page
                      - total
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                      last:
                        type: string
                      prev:
                        type:
                          - string
                          - 'null'
                      next:
                        type:
                          - string
                          - 'null'
                    required:
                      - first
                      - last
                      - prev
                      - next
                required:
                  - message
                  - data
                  - meta
                  - links
      security: []
  '/v1/razas/{raza}/animales':
    get:
      operationId: razas.animales.index
      description: >-
        Retorna el listado paginado de animales pertenecientes a una raza
        específica.

        Valida previamente la existencia de la raza indicada en la ruta.
      summary: Listar animales por raza
      tags:
        - Animal
      parameters:
        - name: raza
          in: path
          required: true
          description: Identificador único de la raza.
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Listado de animales de la raza obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - type: object
                        additionalProperties:
                          type: string
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                      last_page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
                    required:
                      - current_page
                      - last_page
                      - per_page
                      - total
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                      last:
                        type: string
                      prev:
                        type:
                          - string
                          - 'null'
                      next:
                        type:
                          - string
                          - 'null'
                    required:
                      - first
                      - last
                      - prev
                      - next
                required:
                  - message
                  - data
                  - meta
                  - links
      security: []
...
  /v1/potreros:
    get:
      operationId: potreros.index
      description: >-
        Retorna el listado paginado de potreros de la finca con filtros por
        nombre o estado del pasto.
      summary: Listar potreros
      tags:
        - Potrero
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
      security: []
    post:
      operationId: potreros.store
      description: >-
        Crea un potrero en el sistema indicando capacidad, área y estado del
        pasto.
      summary: Registrar un nuevo potrero
      tags:
        - Potrero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorePotreroRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 201
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
  '/v1/potreros/{id}':
    get:
      operationId: potreros.show
      description: Retorna la información de un potrero específico por su ID.
      summary: Obtener detalle de un potrero
      tags:
        - Potrero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
      security: []
    put:
      operationId: potreros.update
      description: Actualiza la información de un potrero existente.
      summary: Actualizar datos de un potrero
      tags:
        - Potrero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePotreroRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: potreros.destroy
      description: >-
        Elimina el potrero indicado. Falla si el potrero tiene animales
        asignados actualmente (RN-02).
      summary: Eliminar un potrero
      tags:
        - Potrero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
      security: []
  /v1/razas:
    get:
      operationId: razas.index
      description: Retorna el listado paginado de razas registradas en el sistema.
      summary: Listar razas
      tags:
        - Raza
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Listado de razas obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - type: object
                        additionalProperties:
                          type: string
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                      last_page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
                    required:
                      - current_page
                      - last_page
                      - per_page
                      - total
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                      last:
                        type: string
                      prev:
                        type:
                          - string
                          - 'null'
                      next:
                        type:
                          - string
                          - 'null'
                    required:
                      - first
                      - last
                      - prev
                      - next
                required:
                  - message
                  - data
                  - meta
                  - links
      security: []
    post:
      operationId: razas.store
      description: Valida los datos y registra una nueva raza de ganado en el sistema.
      summary: Registrar una nueva raza
      tags:
        - Raza
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRazaRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Raza creada correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/RazaResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
  '/v1/razas/{id}':
    get:
      operationId: razas.show
      description: Retorna la información de una raza específica por su ID.
      summary: Obtener detalle de una raza
      tags:
        - Raza
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Raza obtenida correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/RazaResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: razas.update
      description: Actualiza la información de una raza existente por su ID.
      summary: Actualizar una raza
      tags:
        - Raza
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRazaRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Raza actualizada correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/RazaResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: razas.destroy
      description: Elimina del sistema la raza indicada por su ID.
      summary: Eliminar una raza
      tags:
        - Raza
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
components:
  securitySchemes:
    http:
      type: http
      scheme: bearer
  schemas:
    AnimalCollection:
      type: array
      items:
        allOf:
          - $ref: '#/components/schemas/AnimalResource'
          - type: object
            required:
              - raza
              - potrero
      title: AnimalCollection
    AnimalResource:
      type: object
      properties:
        id_animal:
          type: integer
        numero_arete:
          type: string
        sexo:
          type: string
        fecha_nacimiento:
          type:
            - string
            - 'null'
        estado:
          type: string
        raza_id:
          type: integer
        potrero_id:
          type: integer
        raza:
          $ref: '#/components/schemas/RazaResource'
        potrero:
          $ref: '#/components/schemas/PotreroResource'
        pesajes:
          type: array
          items:
            $ref: '#/components/schemas/PesajeResource'
        tratamientos:
          type: array
          items:
            $ref: '#/components/schemas/TratamientoResource'
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - id_animal
        - numero_arete
        - sexo
        - fecha_nacimiento
        - estado
        - raza_id
        - potrero_id
        - created_at
        - updated_at
      title: AnimalResource
...
    PesajeResource:
      type: object
      properties:
        pesaje_id:
          type: integer
        id_animal:
          type: integer
        peso_kg:
          type: number
        fecha_pesaje:
          type:
            - string
            - 'null'
        observaciones:
          type:
            - string
            - 'null'
        animal:
          type: object
          description: 'Resumen del animal, solo si la relación fue cargada.'
          properties:
            id_animal:
              type: integer
            numero_arete:
              type: string
          required:
            - id_animal
            - numero_arete
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - pesaje_id
        - id_animal
        - peso_kg
        - fecha_pesaje
        - observaciones
        - created_at
        - updated_at
      title: PesajeResource
    PotreroResource:
      type: object
      properties:
        potrero_id:
          type: integer
        nombre:
          type: string
        hectareas_de_extension:
          type: number
        capacidad_maxima:
          type: integer
        estado_pasto:
          type:
            - string
            - 'null'
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - potrero_id
        - nombre
        - hectareas_de_extension
        - capacidad_maxima
        - estado_pasto
        - created_at
        - updated_at
      title: PotreroResource
    RazaResource:
      type: object
      properties:
        raza_id:
          type: integer
        nombre:
          type: string
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - raza_id
        - nombre
        - created_at
        - updated_at
      title: RazaResource
...
    StoreAnimalRequest:
      type: object
      properties:
        numero_arete:
          type: string
          maxLength: 50
        raza_id:
          type: integer
        sexo:
          type: string
          enum:
            - Macho
            - Hembra
        fecha_nacimiento:
          type:
            - string
            - 'null'
          format: date-time
        estado:
          type:
            - string
            - 'null'
          maxLength: 50
        potrero_id:
          type: integer
      required:
        - numero_arete
        - raza_id
        - sexo
        - potrero_id
      title: StoreAnimalRequest
...
    StorePotreroRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        hectareas_de_extension:
          type: number
          minimum: 0
          maximum: 999999.99
        capacidad_maxima:
          type: integer
          minimum: 0
        estado_pasto:
          type:
            - string
            - 'null'
          maxLength: 50
      required:
        - nombre
        - hectareas_de_extension
        - capacidad_maxima
      title: StorePotreroRequest
    StoreRazaRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
      required:
        - nombre
      title: StoreRazaRequest
...
    TratamientoResource:
      type: object
      properties:
        tratamiento_id:
          type: integer
        nombre:
          type: string
        descripcion:
          type:
            - string
            - 'null'
        tipo:
          type:
            - string
            - 'null'
        aplicacion:
          type: object
          properties:
            tratamiento_animal_id:
              type: string
            fecha_aplicacion:
              type:
                - string
                - 'null'
            dosis_ml:
              type: string
            observaciones:
              type: string
          required:
            - tratamiento_animal_id
            - fecha_aplicacion
            - dosis_ml
            - observaciones
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - tratamiento_id
        - nombre
        - descripcion
        - tipo
        - created_at
        - updated_at
      title: TratamientoResource
...
    UpdateAnimalRequest:
      type: object
      properties:
        numero_arete:
          type: string
          maxLength: 50
        raza_id:
          type: integer
        sexo:
          type: string
          enum:
            - Macho
            - Hembra
        fecha_nacimiento:
          type:
            - string
            - 'null'
          format: date-time
        estado:
          type:
            - string
            - 'null'
          maxLength: 50
        potrero_id:
          type: integer
      title: UpdateAnimalRequest
...
    UpdatePotreroRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        hectareas_de_extension:
          type: number
          minimum: 0
          maximum: 999999.99
        capacidad_maxima:
          type: integer
          minimum: 0
        estado_pasto:
          type:
            - string
            - 'null'
          maxLength: 50
      title: UpdatePotreroRequest
    UpdateRazaRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
      title: UpdateRazaRequest
...
  responses:
...
    ValidationException:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Errors overview.
              errors:
                type: object
                description: A detailed description of each field that failed validation.
                additionalProperties:
                  type: array
                  items:
                    type: string
            required:
              - message
              - errors